+++ test/functional-tests/features/conditionalLogic.feature
@@ -1,0 +1,55 @@
+@functional @conditional @showandhide
+Feature: Set of scenarios testing show & hide functionality and conditional logic
+
+  Background:
+    Given I have logged in
+
+  Scenario: field NOT visible when show & hide condition is NOT met
+    And a case type containing conditionals exists
+    When I enter 'dontshowmethemoney' into the 'text' field
+    Then the field with label 'Text Field 2' is not visible
+
+  Scenario: field visible when show & hide condition is met
+    And a case type containing conditionals exists
+    When I enter 'showmethemoney' into the 'text' field
+    Then the field with label 'Text Field 2' is visible
+
+  Scenario: page NOT visible when show & hide condition is NOT met
+    And a case type containing conditionals exists
+    And I enter 'dontshowmethemoney' into the 'text' field
+    When I click the Continue button
+    Then I will not be on the 'Conditional Page 2 - Create a case' page
+
+  Scenario: page visible when show & hide condition is met
+    And a case type containing conditionals exists
+    And I enter 'showmethemoney' into the 'text' field
+    When I click the Continue button
+    Then I will be on the 'Conditional Page 2 - Create a case' page
+
+
+  Scenario: tab visible when show & hide condition is met
+    And a case type containing conditionals exists
+    And I enter 'showmethemoney' into the 'text' field
+    When I Submit the case
+    Then the following tabs will be visible:
+      | History             |
+      | Conditional Tab 1   |
+
+  Scenario: tab NOT visible when show & hide condition is NOT met
+    And a case type containing conditionals exists
+    And I enter 'dontshowmethemoney' into the 'text' field
+    When I Submit the case
+    Then the following tabs will be visible:
+      | History |
+
+  Scenario: field in tab visible when show & hide condition is met
+    And a case type containing conditionals exists
+    When I meet the condition for showing the field in the tab
+    Then the 'Text Field 3' field will be visible on the 'Conditional Tab 1' tab
+
+  Scenario: field in tab NOT visible when show & hide condition is NOT met
+    And a case type containing conditionals exists
+    When I do NOT meet the condition for showing the field in the tab
+    Then the 'Text Field 3' field will NOT be visible on the 'Conditional Tab 1' tab
+
+
